Added a test case for simulate_period.
Created a test case for the simulation of a single period. It seems to hit most of the code. Further testing is probably needed to ensure the correct working of the overal simulation flow. Several bugs have been fixed during the creation of this test that were the result of unexpected behaviour and unexpected interactions.