Add the metrics module
Metrics will be emitted in a JSON compatible format

The simulation can output events using `log_event()`.
By default these metrics won't be logged, the using application can call `setup_metrics()` to initilize the logger.
If a filename is passed to setup_metircs, the metrics will be written to that logfile instead of stdout.