Let the metrics file output be valid JSON
To allow parsing of an entire metrics file output ensure the files created are valid JSON.
The change adds the JsonRotatingFileHandler as the default metrics handler.

This handler assumes the logs created are valid JSON on their own and adds them to the logfile as an item in a root list.