fix(node): Ensure single-echelon nodes get their `llc` set
Nodes without predecessors did not get their llc set during SupplyChain init which resulted in a llc of -1.

llc's with -1 were not included in the control loop, resulting in missing Receipts in the pipeline during simulation.

This was a mistaken adaption of an optimisation of the set_llc method, it should have been the successors == 0.
As we don't store the successors for a Node this optimisation can't currently be done during `_set_llc`