feat(supplychain): Add `Supplychain.from_json()` and `SupplyChain().to_json()`
SupplyChain.from_json() allows building a SupplyChain from a JSON file.
It takes a JSON encoded string or a path to a JSON file as input and will return a SupplyChain instance.

SupplyChain().to_json() will export the SupplyChain to a JSON encoded string.
It takes an optional file path to write the JSON data to.