Setup of loop functionality
It is now possible to add the 'loops' argument to simulator.run. This argument makes it possible to duplicate a simulation while retaining the state from one loop to the next. This greatly simplifies cold starts, removing the need to duplicate information such as sales order lines, reducing the size of the .json files as well the potential for errors such as an off by one problem. Also included is a refactor of the SupplyChain class which is moved into its own file.